requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
PySide6>=6.5.0
pytest>=7.3.0
pyinstaller>=6.0.0
Pillow>=10.0.0
